Update how we get latest story ID
Hacker News used to post 3 different posts at the beginning of the
month, and the assumption was that one of these posts was the latest
'Who is Hiring?' post. Recently they stopped submitting posts for
Freelancers which was causing issues with my sync process.

Instead of checking for the first 2 submissions by Hacker News, I will
continue to check the first 3 but added some additional conditions to
determine when we should continue to use the saved existing story vs
fetching the most recent one.